==> CarbChem/3.GAM_BAMmodels_filtereddata.Rmd <==
---
title: "BAM_Models"
author: "Ana Palacio-Castro"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
    fig_height: 8
    fig_width: 8
    df_print: paged
    toc: yes
    toc_float: true
editor_options:
  chunk_output_type: console
---


```{r setup, include = FALSE}
  knitr::opts_chunk$set(warning = FALSE, 
                        message = FALSE,
                        #options(knitr.kable.NA = ''),
                        fig.align = "center")
```


```{r ibraries, include=FALSE}
library('mgcv')
library('ggplot2')
library('viridis')
theme_set(theme_bw())
library('gganimate')
library(plyr)
library(dplyr)
```

# Data 

```{r, include=FALSE}
# 1. Get data
WS.data<-read.csv("FLK.data_filtered2.csv", header = T)

# Separate CalVal and Walton Smith data
summary(WS.data)
summary(WS.data$Extreme)

```

## Create time variables for the GAM

```{r}
# WS.data <- transform(WS.data,
#                        datetime = as.POSIXct(paste(ESTDate, EST.Time),
#                                              format = '%m/%d/%Y %I:%M:%S %p', tz = "EST"))
# summary(WS.data$datetime)
# WS.data <- transform(WS.data,
#                        SiteID = factor(SiteID),
#                        MoY = as.numeric(format(datetime, format = '%m')),
#                        #DoY = as.numeric(format(datetime, format = '%j')),
#                        ToD = as.numeric(format(datetime, format = '%H')) +
#                          (as.numeric(format(datetime, format = '%M')) / 60))
```


## Subset FLK data by zone to use them as endpoints

```{r}
  WS.open<-subset(WS.data, WS.data$Zone=="Offshore"|WS.data$Zone=="Oceanic")
  WS.open<-subset(WS.open, WS.open$Extreme=="Normal")
  WS.open$Date<-as.Date(WS.open$Date, "%Y-%m-%d")
 
  WS.reef<-subset(WS.data, WS.data$Zone=="Inshore")
  WS.reef<-subset(WS.reef, WS.reef$Extreme=="Normal")
  WS.reef$Date<-as.Date(WS.reef$Date, "%Y-%m-%d")
```

* Get the oceanic / offshore mean values per monty-year and substract them from the reef data

```{r}

# # Reef data
# nTA.MY.open <- ddply (WS.open, .(Sub_region, MY),summarise,
#                 OnTA_my = mean (nTA, na.rm = T),
#                 #sd = sd (nTA, na.rm = T),
#                 number = n())
# 
# nDIC.MY.open <- ddply (WS.open, .(Sub_region, MY),summarise,
#                 OnDIC_my = mean (nDIC, na.rm = T),
#                 #sd = sd (nDIC, na.rm = T),
#                 number = n())
# 
# # ONly reef
#   WS.reef<-join(WS.reef,nTA.MY.open[,-4], 
#                    type = "left", by=c("MY", "Sub_region"))
#   
#   WS.reef<-join(WS.reef, nDIC.MY.open, 
#                    type = "left", by=c("MY", "Sub_region"))
#   
#   # Delta nTA
#   WS.reef$dnTA_MY<-WS.reef$nTA-WS.reef$OnTA_my
#   
#   # Delta nDIC
#   WS.reef$dnDIC_MY<-WS.reef$nDIC-WS.reef$OnDIC_my
# 
# 
# # all WS data
#   WS.data<-join(WS.data,nTA.MY.open[,-4], 
#                    type = "left", by=c("MY", "Sub_region"))
#   
#   WS.data<-join(WS.data, nDIC.MY.open, 
#                    type = "left", by=c("MY", "Sub_region"))
#   
#   # Delta nTA
#   WS.data$dnTA_MY<-WS.data$nTA-WS.data$OnTA_my
#   
#   # Delta nDIC
#   WS.data$dnDIC_MY<-WS.data$nDIC-WS.data$OnDIC_my

```

**Quick plote / note about Time of the day:** Either hour of the day does not matter for temperature (lots of circulation?) or times are not very accurate (Waton Smith UTC reporting not always followed?) 

## Check undesirable patters in the data

### Time of the day

```{r}
# ToDay<-ggplot(WS.data, aes(ToD, Aragonite_Sat_W)) +
#   geom_smooth() +
#   theme(panel.border = element_blank(),
#         panel.background = element_blank(),
#         panel.grid.minor = element_line(colour = "grey90"),
#         panel.grid.major = element_line(colour = "grey90"),
#         panel.grid.major.x = element_line(colour = "grey90"),
#         axis.text = element_text(size = 10),
#         axis.title = element_text(size = 12, face = "bold")) + facet_wrap(~MoY)
# ToDay + geom_point(aes(colour=Zone))
```

### Time of the day

```{r}
# MoYear<-ggplot(WS.data, aes(MoY, Aragonite_Sat_W)) +
#   geom_smooth() +
#   theme(panel.border = element_blank(),
#         panel.background = element_blank(),
#         panel.grid.minor = element_line(colour = "grey90"),
#         panel.grid.major = element_line(colour = "grey90"),
#         panel.grid.major.x = element_line(colour = "grey90"),
#         axis.text = element_text(size = 10),
#         axis.title = element_text(size = 12, face = "bold"))
# MoYear + geom_point(aes(colour=Zone))+facet_wrap(~Sub_region)
```

# Generalized additive models (GAMs) for fitting irregular time series

## Notes

The principle behind GAMs is similar to that of regression, except that instead of summing effects of individual predictors, GAMs are a sum of smooth functions. Functions allow us to model more complex patterns, and they can be averaged to obtain smoothed curves that are more generalizable. Because GAMs are based on functions rather than variables, they are not restricted by the linearity assumption in regression that requires predictor and outcome variables to move in a straight line. Furthermore, unlike in neural networks, we can isolate and study effects of individual functions in a GAM on resulting predictions.

* model the response variable by independent variables, which are in the form of some smooth functions
* Smooth functions are also called splines. Smoothing splines are real functions that are piecewise-defined by polynomial functions (basis functions). The places, where the polynomial pieces connect are called knots. In GAMs, penalized regression splines are used in order to regularize the smoothness of a spline.


There are several smoothing bases b (splines) which are suitable for regression:
* thin plate regression
* cubic regression
* cyclic cubic regression 
* P-splines

bs="tp": Low rank isotropic smoothers of any number of covariates. 

* Isotropic = rotation of the covariate co-ordinate system will not change the result of smoothing
* Low rank = they have far fewer coefficients than there are data to smooth. They are reduced rank versions of the thin plate splines and use the thin plate spline penalty. 
* They are the default smooth for s terms because there is a defined sense in which they are the optimal smoother of any given basis dimension/rank (Wood, 2003).
* Thin plate regression splines do not have ‘knots’ (at least not in any conventional sense): a truncated eigen-decomposition is used to achieve the rank reduction. See tprs for further details.

bs="ts" is as "tp" but with a modification to the smoothing penalty, so that the null space is also penalized slightly and the whole term can therefore be shrunk to zero.

Duchon splines: bs="ds". LONGITUD AND LATITUD
* Thin plate splines. For any given number of covariates they allow lower orders of derivative in the penalty than thin plate splines (and hence a smaller null space). See Duchon.spline for further details.


Smoothing parameter (λ) and the number of basis dimensions (i.e. degrees of freedom):
* GAM uses Generalized Cross Validation score (GCV) to minimize vg
* when lambda is near 1 then spline will be over-smoothed
* when lambda is near zero than spline isn’t penalized, so the method behaves like a classical OLS.
* With a number of basis dimensions (estimated degrees of freedom), it is opposite. Higher dimension implies that fit will be less smoothed (overfit), on the other side lower dimensions implies more smoothed behavior of fitted values.

Interactions:
* Multiplication of two independent variables
* smoothed function to one variable
* same smoothed function for both variables: 
* tensor product interactions: different smoothing bases for variables and penalize it in two (when we do interactions of two independent variables) different ways: 



# GAM models with bam

All these models below include location as coordinates and not as categories in Zone (In- off-shore. etc), UK, MK and LK. That might be better or not?

Number of years, monts (of the year) and time of the day (useless) in the data

```{r model parameters}

n_Year <- unique(WS.data[, "Year"])
length(n_Year) 

n_MoY <- unique(WS.data[, "MoY"])
length(n_MoY)

n_ToD <- unique(WS.data [, "ToD"])
length(n_ToD)

N <- nrow(WS.data) # number of observations in the train set
period<-N/(length(n_Year))/12
period<-round(period)
window <- N / period # number of days in the train set

knots <- list(DoY = c(0.5, 366.5))
M <- list(c(1, 0.5), NA)

```

# All data - BAM: Generalized additive models for very large datasets 

### Temperature

```{r}
# 1. Model
m_0 <- bam(Temperature_C ~ 
           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(50, 12)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(50, 6)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(50, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

#g_0 <- gam(Temperature_C ~
#           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
#           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
#           s(Year, k = 5) +
#           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
#           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
#           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','tp'), 
#              m = M, k = c(50, 12)) +
#           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
#              m = M, k = c(50, 6)) +
#            ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
#              m = M, k = c(50, 5)),
#data = WS.data, method = 'REML', knots = knots, nthreads = 4, discrete = TRUE)
##summary(g_0)
#plot(g_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Temperature_C ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(50, 12)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(50, 6)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)
plot(m_1, pages = 1, scheme = 2, shade = TRUE)

AIC(m_0, m_1)
gam.check(m_0)
gam.check(m_1)

#summary(mA)$s.table
m_1$optimizer
summary(m_1)$sp.criterion

# 2. Effects
#plot(m, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_1, pages = 1, scheme = 2, shade = TRUE)
#plot(m_1, shade = TRUE)

#ToD and MoY
vis.gam(m_1, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
vis.gam(m_1, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_1, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1, 3, 5, 7, 9, 11),
                          Year = seq(min(Year), max(Year), length = 20),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_1, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(Temperature ~ (degree * C)))

```

## Omega

```{r}
# 1. Model

m_0 <- bam(Aragonite_Sat_W  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Aragonite_Sat_W  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)

m_2 <- bam(Aragonite_Sat_W  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_2)

AIC(m_0, m_1, m_2) 

m_omega <-m_2

gam.check(m_omega)

# 2. Effects
#plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_omega, pages = 1, scheme = 2, shade = TRUE)



# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_omega, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("Aragonite saturation state "~Omega), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), length = 12),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_omega, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

```

## pCO2

```{r}
# 1. Model

mpCO2_0 <- bam(pCO2  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_0)

mpCO2_1 <- bam(pCO2  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_1)

mpCO2_2 <- bam(pCO2  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_2)

AIC(mpCO2_0, mpCO2_1, mpCO2_2) 

m_pCO2 <-mpCO2_1

gam.check(m_pCO2)

#ToD and MoY
vis.gam(m_pCO2, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
vis.gam(m_pCO2, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 2. Effects
#plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pCO2, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("pCO2"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_pCO2, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pCO2))

```

**RESULTS: **
* Seasonality remains the same, but pCO2 values are higher. 
* Differences among years are bigger in summer-fall and smaller in winter-spring

## pH

```{r}
# 1. Model

mpH_0 <- bam(pH  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_0)

mpH_1 <- bam(pH  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_1)

mpH_2 <- bam(pH  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_2)

AIC(mpH_0, mpH_1,mpH_2) 

m_pH <-mpH_1

gam.check(m_pH)

#ToD and MoY
vis.gam(m_pH, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
vis.gam(m_pH, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)

# 2. Effects
#plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_pH, pages = 1, scheme = 2, shade = TRUE)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pH, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("pH"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by =1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_pH, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pH))

```



**RESULTS: **
* Seasonality remains very similar, with overall decline in pH. Small displacement of minimun pH from Oct to Sep? 
* Differences among years are bigger in summer-fall and smaller in winter-spring


## nDIC

```{r}
# 1. Model

mnDIC_0 <- bam(nDIC  ~
          s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_0)

mnDIC_1 <- bam(nDIC  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_1)

mnDIC_2 <- bam(nDIC  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_2)

AIC(mnDIC_0, mnDIC_1, mnDIC_2) 

m_DIC <-mnDIC_0

gam.check(m_DIC)

# 2. Effects
  #plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_DIC, pages = 1, scheme = 2, shade = TRUE)
  
  #ToD and MoY
  vis.gam(m_DIC, n.grid = 50, theta = 35, phi = 32, zlab = "",
          ticktype = "detailed", color = "topo", main = "t2(D, W)")
  vis.gam(m_DIC, main = "t2(D, W)", plot.type = "contour",
          color = "terrain", contour.col = "black", lwd = 2)

# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_DIC, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nDIC"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_DIC, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(nDIC))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(nDIC))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(nDIC))

```

**RESULTS: **
* ?????


## nTA

```{r}
# 1. Model

mnTA_0 <- bam(nTA  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_0)

mnTA_1 <- bam(nTA  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.data, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_1)

AIC(mnTA_0, mnTA_1) 

m_TA <-mnTA_0

gam.check(m_TA)

# 2. Effects
  #plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_TA, pages = 1, scheme = 2, shade = TRUE)
  
  #ToD and MoY
  vis.gam(m_TA, n.grid = 50, theta = 35, phi = 32, zlab = "",
          ticktype = "detailed", color = "topo", main = "t2(D, W)")
  vis.gam(m_TA, main = "t2(D, W)", plot.type = "contour",
          color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_TA, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.data$Longitude, WS.data$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nTA"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.data,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_TA, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(nTA))

```

# Ofshore+Oceanic data 

### Temperature

```{r}
# 1. Model
m_0 <- bam(Temperature_C ~ Sub_region +
           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
            s(ToD, by=Sub_region, m=1) + # Time of the day//added bs='cr'   
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(50, 12)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(50, 6)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(50, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

#g_0 <- gam(Temperature_C ~
#           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
#           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
#           s(Year, k = 5) +
#           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
#           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
#           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','tp'), 
#              m = M, k = c(50, 12)) +
#           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
#              m = M, k = c(50, 6)) +
#            ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
#              m = M, k = c(50, 5)),
#data = WS.data, method = 'REML', knots = knots, nthreads = 4, discrete = TRUE)
##summary(g_0)
#plot(g_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Temperature_C ~ 
           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
           #s(ToD, by=Sub_region, m=1) + # Time of the day//added bs='cr'   
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           #s(MoY, by=Sub_region, m=1) + # Time of the day//added bs='cr'   
           s(Year, k = 5) +
          #  s(Year, by=Sub_region, m=1)+
           #s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)),# +
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
          #   m = M, k = c(50, 12)) +
           #ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              #m = M, k = c(50, 6)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)
plot(m_1, pages = 1, scheme = 2, shade = TRUE)

AIC(m_0, m_1)
par(mfrow=c(2,2))
gam.check(m_0)
gam.check(m_1)
par(mfrow=c(1,1))

#summary(mA)$s.table
m_1$optimizer
summary(m_1)$sp.criterion

# 2. Effects
#plot(m, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_1, pages = 1, scheme = 2, shade = TRUE)
#plot(m_1, shade = TRUE)

#ToD and MoY
vis.gam(m_1, view=c("Year","MoY"), n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")

vis.gam(m_1, view=c("Year","MoY"), main = "Temperature", plot.type = "contour",
            color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_1, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by =1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_1, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(Temperature ~ (degree * C)))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(Temperature ~ (degree * C)))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(Temperature ~ (degree * C)))


```

## Omega

```{r}
# 1. Model

m_0 <- bam(Aragonite_Sat_W  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Aragonite_Sat_W  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)),
      data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)

AIC(m_0, m_1) 

m_omega <-m_1

par(mfrow=c(2,2))
gam.check(m_omega)
par(mfrow=c(1,1))

# 2. Effects
  #plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_omega, pages = 1, scheme = 2, shade = TRUE)
  
  #ToD and MoY
  vis.gam(m_omega, n.grid = 50, theta = 35, phi = 32, zlab = "",
          ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(m_omega, view=c("Year","MoY"), main = "Omega", plot.type = "contour",
          color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_omega, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("Aragonite saturation state "~Omega), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_omega, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper,  fill = factor(Year)), alpha = 0.5) +
  geom_line()+
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

```

## pCO2

```{r}
# 1. Model

mpCO2_0 <- bam(pCO2_uatm  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_0)

mpCO2_1 <- bam(pCO2_uatm  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)), 
      data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_1)

AIC(mpCO2_0, mpCO2_1) 

m_pCO2 <-mpCO2_1
gam.check(m_pCO2)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(m_pCO2, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(m_pCO2, view=c("Year","MoY"), main = "pCO2", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pCO2, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("pCO2"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1)))
fit <- data.frame(predict(m_pCO2, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pCO2))

```

## pH

```{r}
# 1. Model

mpH_0 <- bam(pH_calculated  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_0)

mpH_1 <- bam(pH_calculated  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)),
           #ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
            #  m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_1)

mpH_2 <- bam(pH_calculated  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
           
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_2)

AIC(mpH_0, mpH_1,mpH_2) 

m_pH <-mpH_1

gam.check(m_pH)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_pH, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(m_pH, view=c("Year","MoY"), n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(m_pH, view=c("Year","MoY"), main = "pH", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)

# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pH, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("pH"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by =1),
                          Year = seq(min(Year), max(Year), by =1)))
                          #Longitude = -80.4,
                          #Latitude  = 25.1))
fit <- data.frame(predict(m_pH, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pH))

```



**RESULTS: **

## nDIC

```{r}
# 1. Model

mnDIC_0 <- bam(nDIC  ~
          s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_0)

mnDIC_1 <- bam(nDIC  ~
          s(ToD, k = 12, bs = 'cc') + # Time of the day
          s(MoY, k = 6, bs = 'cc') +  # Month of the year?
          s(Year, k = 5) +
          s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
          ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
          ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
          ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_1)

mnDIC_2 <- bam(nDIC  ~
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
              s(MoY, k = 6, bs = 'cc') +  # Month of the year?
          s(Year, k = 5) +
          s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
          ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
          ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(10, 5)) +
          ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_2)

AIC(mnDIC_0, mnDIC_1, mnDIC_2) 

m_DIC <-mnDIC_2

gam.check(m_DIC)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_DIC, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(m_DIC, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  vis.gam(m_DIC, main = "t2(W, D)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)

# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_DIC, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nDIC"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_DIC, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(nDIC))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(nDIC))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(nDIC))

```

**RESULTS: **
* ?????


## nTA

```{r}
# 1. Model

mnTA_0 <- bam(nTA  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_0)

mnTA_1 <- bam(nTA  ~
           #s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           #ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
           #   m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.open, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_1)

AIC(mnTA_0, mnTA_1) 

m_TA <-mnTA_0
gam.check(m_TA)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_TA, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(m_TA, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  vis.gam(m_TA, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_TA, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nTA"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.open,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_TA, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(nTA))

```

# Reef 

### Temperature

```{r}
# 1. Model
m_0 <- bam(Temperature_C ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(50, 12)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(50, 6)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(50, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Temperature_C ~
          s(ToD, k = 12, bs = 'cc') + # Time of the day//added bs='cr'
          s(MoY, k = 6, bs = 'cc') +  # Month of the year?
          s(Year, k = 5) +
          s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
          ti(MoY, Year, bs = c('cc', 'tp'), k = c(6, 5)) +
          ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
            m = M, k = c(50, 12)) +
          ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(50, 6)), data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)
plot(m_1, pages = 1, scheme = 2, shade = TRUE)

AIC(m_0, m_1)
par(mfrow=c(2,2))
gam.check(m_0)
gam.check(m_1)
par(mfrow=c(1,1))

m_temp<-m_0
#summary(mA)$s.table
#m_temp$optimizer
#summary(m_temp)$sp.criterion

# 2. Effects
#plot(m, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_temp, pages = 1, scheme = 2, shade = TRUE)
#plot(m_1, shade = TRUE)

#ToD and MoY
vis.gam(m_temp, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
vis.gam(m_1, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_temp, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.reef$Longitude, WS.reef$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by =1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_temp, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(Temperature ~ (degree * C)))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(Temperature ~ (degree * C)))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(Temperature ~ (degree * C)))


```

## Omega

```{r}
# 1. Model

m_0 <- bam(Aragonite_Sat_W  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_0)
plot(m_0, pages = 1, scheme = 2, shade = TRUE)

m_1 <- bam(Aragonite_Sat_W  ~
           #s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           #s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)),
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
             # m = M, k = c(10, 5)) +
           #ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
            #  m = M, k = c(15, 5)),
      data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(m_1)

AIC(m_0, m_1) 

m_omega <-m_1
summary(m_omega)

# 2. Effects
#plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_omega, pages = 1, scheme = 2, shade = TRUE)
 
#ToD and MoY
  vis.gam(m_omega, n.grid = 50, theta = 35, phi = 32, zlab = "",
          ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(m_omega, view=c("Year","MoY"), main = "Omega", plot.type = "contour",
          color = "terrain", contour.col = "black", lwd = 2)



# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_omega, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("Aragonite saturation state "~Omega), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(m_omega, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper,  fill = factor(Year)), alpha = 0.5) +
  geom_line()+
  labs(x = NULL, y = expression("Aragonite saturation state "~Omega))

```

## pCO2

```{r}
# 1. Model

mpCO2_0 <- bam(pCO2_uatm  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_0)

mpCO2_1 <- bam(pCO2_uatm  ~
          #s(ToD, k = 12, bs = 'cc') + # Time of the day
          s(Year, k = 5) +
          s(MoY, k = 6, bs = 'cc') +  # Month of the year?
          #s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
          ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)), 
          # ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
          #    m = M, k = c(10, 5)) +
           #ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
          #    m = M, k = c(15, 5)),
      data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpCO2_1)

AIC(mpCO2_0, mpCO2_1) 

m_pCO2 <-mpCO2_1

# 2. Effects
#plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE)
#ToD and MoY
vis.gam(m_pCO2, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
vis.gam(m_pCO2, main = "pCO2", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = 1,
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pCO2, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.reef$Longitude, WS.reef$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_wrap(~ Year, ncol = 4) +
  scale_fill_viridis(name = expression("pCO2"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = 24.7,
                          Latitude  = -80.83))
fit <- data.frame(predict(m_pCO2, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pCO2))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pCO2))

```

## pH

```{r}
# 1. Model

mpH_0 <- bam(pH  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_0)

mpH_1 <- bam(pH_calculated  ~
           #s(ToD, k = 12, bs = 'cc') + # Time of the day
          s(Year, k = 5) +
          s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           #s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)),
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
          #    m = M, k = c(10, 5)) +
           #ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
          #    m = M, k = c(15, 5)) ,
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mpH_1)

AIC(mpH_0, mpH_1) 

m_pH <-mpH_1

gam.check(m_pH)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(m_pH, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(m_pH, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(m_pH, main = "pH", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(m_pH, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.reef$Longitude, WS.reef$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("pH"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by =1),
                          Year = seq(min(Year), max(Year), by =1),
                          Longitude = -80.56,
                          Latitude  = 24.9))
fit <- data.frame(predict(m_pH, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(pH))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(pH))

```



**RESULTS: **

## nDIC

```{r}
# 1. Model

mnDIC_0 <- bam(dnDIC_MY  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_0)

mnDIC_1 <- bam(dnDIC_MY  ~
          #s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           #s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
            #  m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_1)

mnDIC_2 <- bam(dnDIC_MY  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           #s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
            #  m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnDIC_2)

AIC(mnDIC_0, mnDIC_1, mnDIC_2) 

mn_DIC <-mnDIC_1

gam.check(mn_DIC)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(mn_DIC, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(mn_DIC, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(mn_DIC, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(mn_DIC, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.open$Longitude, WS.open$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nDIC"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = c(1,2, 3, 4, 5, 6, 7,8, 9,10, 11, 12),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(mn_DIC, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(Delta~"nDIC"))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(Delta~"nDIC"))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(Delta~"nDIC"))

```

**RESULTS: **
* ?????


## nTA

```{r}
# 1. Model

mnTA_0 <- bam(dnTA_MY  ~
           s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
              m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_0)

mnTA_1 <- bam(dnTA_MY  ~
           # s(ToD, k = 12, bs = 'cc') + # Time of the day
           s(MoY, k = 6, bs = 'cc') +  # Month of the year?
           #s(Year, k = 5) +
           s(Longitude, Latitude, k = 50, bs = 'ds', m = c(1, 0.5)) +
           ti(MoY, Year, bs = c('cc', 'tp'), k = c(5, 4)) +
           #ti(Longitude, Latitude, ToD, d = c(2,1), bs = c('ds','cc'), 
          #    m = M, k = c(10, 5)) +
           ti(Longitude, Latitude, MoY, d = c(2,1), bs = c('ds','cc'),
              m = M, k = c(15, 5)) +
           ti(Longitude, Latitude, Year, d = c(2,1), bs = c('ds','tp'), 
              m = M, k = c(15, 5)),
data = WS.reef, method = 'fREML', knots = knots, nthreads = 4, discrete = TRUE)
summary(mnTA_1)

AIC(mnTA_0, mnTA_1) 

mn_TA <-mnTA_0

gam.check(mn_TA)

# 2. Effects
  #plot(m_omega, pages = 1, scheme = 2, shade = TRUE, scale = 0)
  plot(mn_TA, pages = 1, scheme = 2, shade = TRUE)
  
  vis.gam(mn_TA, n.grid = 50, theta = 35, phi = 32, zlab = "",
        ticktype = "detailed", color = "topo", main = "t2(D, W)")
  
  vis.gam(mn_TA, main = "t2(D, W)", plot.type = "contour",
        color = "terrain", contour.col = "black", lwd = 2)

# 2. Effects
#plot(m_pCO2, pages = 1, scheme = 2, shade = TRUE, scale = 0)
plot(mn_TA, pages = 1, scheme = 2, shade = TRUE)


# 3. Predictions 

# 3.1 spatial pattern and how they vary over the years -> predict from the model for a 100x100 grid over the spatial domain, at midday on month 1 of each year:
pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = c(1,9),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = seq(min(Longitude), max(Longitude), length = 100),
                          Latitude  = seq(min(Latitude), max(Latitude), length = 100)))
fit <- predict(mn_TA, pdata)

ind <- exclude.too.far(pdata$Longitude, pdata$Latitude,
                       WS.reef$Longitude, WS.reef$Latitude, dist = 0.1)
fit[ind] <- NA
pred <- cbind(pdata, Fitted = fit)

ggplot(pred, aes(x = Longitude, y = Latitude)) +
  geom_raster(aes(fill = Fitted)) + facet_grid(MoY~ Year) +
  scale_fill_viridis(name = expression("nTA"), option = 'plasma',
                     na.value = 'transparent') +
  coord_quickmap() +
  theme(legend.position = 'top', legend.key.width = unit(2, 'cm'))

#p<-ggplot(pred, aes(x = Longitude, y = Latitude, frame = Year)) +
#  geom_raster(aes(fill = Fitted)) +
#  scale_fill_viridis(name = expression(degree*C), option = 'plasma',
#                     na.value = 'transparent') +
#  coord_quickmap() +
#  theme(legend.position = 'top', legend.key.width = unit(1, 'cm'))+
#  labs(title = 'Year: {frame_time}', x = 'Longitude', y = 'Latitude') +
#  transition_time(year) +
#  ease_aes('linear')

# animate(p, renderer = ffmpeg_renderer())


# 3.2 Long-term trends in more detail ->  predict for particular stations and months 
# chose (midday, on month of year 1, 4, 7, and 10, at the selected station, and evaluating the long-term trend at 30 equally spaced values)

pdata <- with(WS.reef,
              expand.grid(ToD = 12,
                          MoY = seq(1, 12, by = 1),
                          Year = seq(min(Year), max(Year), by = 1),
                          Longitude = -80.4,
                          Latitude  = 25.1))
fit <- data.frame(predict(mn_TA, newdata = pdata, se.fit = TRUE))
fit <- transform(fit, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
pred <- cbind(pdata, fit)

ggplot(pred, aes(x = Year, y = fit, group = factor(MoY))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ MoY) +
  labs(x = NULL, y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, group = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line() + facet_wrap(~ Year) +
  labs(x = "Month", y = expression(nTA))

ggplot(pred, aes(x = MoY, y = fit, colour = factor(Year))) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = 'grey', alpha = 0.5) +
  geom_line()+
  labs(x = "Month", y = expression(nTA))

```
